refactor: [US-048b] - Move static-page views into core/views/pages.py
