+++ scripts/ralph/prd.json
@@ -839,7 +839,7 @@
         "`python manage.py check` passes; URL count check matches.",
         "`manage.py test` passes."
       ],
-      "passes": false,
+      "passes": true,
       "notes": "",
       "priority": 57
     },
